+++ src/main/java/com/mobble/mobbleserver/refactor/notification/core/service/NotificationService.java
@@ -4,13 +4,13 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.mobble.mobbleserver.domain.member.Member;
 import com.mobble.mobbleserver.infrastructure.persistence.member.JpaMemberRepository;
 import com.mobble.mobbleserver.domain.notification.core.Notification;
-import com.mobble.mobbleserver.refactor.notification.core.repository.NotificationRepository;
+import com.mobble.mobbleserver.infrastructure.persistence.notification.core.NotificationRepository;
 import com.mobble.mobbleserver.domain.notification.device.DeviceToken;
-import com.mobble.mobbleserver.refactor.notification.device.repository.DeviceTokenRepository;
+import com.mobble.mobbleserver.infrastructure.persistence.notification.device.DeviceTokenRepository;
 import com.mobble.mobbleserver.domain.notification.outbox.PushOutbox;
-import com.mobble.mobbleserver.refactor.notification.outbox.repository.PushOutboxRepository;
+import com.mobble.mobbleserver.infrastructure.persistence.notification.outbox.PushOutboxRepository;
 import com.mobble.mobbleserver.domain.notification.setting.NotificationSetting;
-import com.mobble.mobbleserver.refactor.notification.setting.repository.NotificationSettingRepository;
+import com.mobble.mobbleserver.infrastructure.persistence.notification.setting.NotificationSettingRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.data.domain.PageRequest;
