+++ src/main/java/com/mobble/mobbleserver/refactor/notification/outbox/worker/PushProcessor.java
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.mobble.mobbleserver.config.push.PushClient;
 import com.mobble.mobbleserver.domain.notification.outbox.PushOutbox;
 import com.mobble.mobbleserver.domain.notification.outbox.Status;
-import com.mobble.mobbleserver.refactor.notification.outbox.repository.PushOutboxRepository;
+import com.mobble.mobbleserver.infrastructure.persistence.notification.outbox.PushOutboxRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.PageRequest;
